pyproject.toml
[project]
name = "project-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "contextily>=1.6.2",
    "db-dtypes>=1.4.3",
    "dvc>=3.60.1",
    "geopandas>=1.1.1",
    "google-cloud>=0.34.0",
    "google-cloud-bigquery>=3.34.0",
    "google-cloud-bigquery-storage>=2.32.0",
    "google-cloud-storage>=3.1.1",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "linearmodels>=6.1",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "patsy>=1.0.1",
    "polars>=1.31.0",
    "pyarrow>=20.0.0",
    "pyyaml>=6.0.2",
    "scipy>=1.16.0",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.4",
]
